app/requirements.txt
alembic~=1.14.1
bcrypt==4.0.1
cloudinary
fastapi~=0.115.8
fastapi_jwt_auth
fastapi_mail
google-cloud-storage
jinja2~=3.1.5
inflect~=7.5.0
pandas
passlib[bcrypt]~=1.7.4
pathlib~=1.0.1
pillow
psycopg2-binary
pydantic~=2.10.6
pydantic-settings~=2.7.1
python-dateutil
#python-dotenv
python-multipart
python-jose[cryptography]~=3.3.0
qrcode
starlette~=0.45.3
sqlalchemy~=2.0.38
uuid~=1.30
uvicorn~=0.34.0
cloud-sql-python-connector==1.2.4
functions-framework==3.3.0
pg8000==1.29.8
python-Levenshtein
fastapi_limiter
slowapi~=0.1.9
certbot
asyncpg
httpx~=0.28.1
protobuf~=5.29.3
requests~=2.32.3